[libc++] Avoid including <format> code in <optional>
